+++ ref-exchange/src/lib.rs
@@ -1,5 +1,6 @@
 use std::convert::TryInto;
 use std::fmt;
+use std::collections::{HashMap, HashSet};
 
 use near_contract_standards::storage_management::{
     StorageBalance, StorageBalanceBounds, StorageManagement,
@@ -22,7 +23,7 @@ use crate::pool::Pool;
 use crate::simple_pool::SimplePool;
 use crate::stable_swap::StableSwapPool;
 use crate::rated_swap::{RatedSwapPool, rate::{RateTrait, global_get_rate, global_set_rate}};
-use crate::utils::{check_token_duplicates, TokenCache};
+use crate::utils::{check_token_duplicates, pair_rated_price_to_vec_u8, TokenCache};
 pub use crate::custom_keys::*;
 pub use crate::views::{PoolInfo, ShadowRecordInfo, RatedPoolInfo, StablePoolInfo, ContractMetadata, RatedTokenInfo, AddLiquidityPrediction, RefStorageState};
 pub use crate::token_receiver::AddLiquidityInfo;
@@ -115,6 +116,7 @@ pub struct Contract {
     unit_share_cumulative_infos: UnorderedMap<u64, VUnitShareCumulativeInfo>,
 
     wnear_id: Option<AccountId>,
+    auto_whitelisted_postfix: HashSet<String>
 }
 
 #[near_bindgen]
@@ -135,7 +137,8 @@ impl Contract {
             referrals: UnorderedMap::new(StorageKey::Referral),
             cumulative_info_record_interval_sec: 12 * 60, // 12 min
             unit_share_cumulative_infos: UnorderedMap::new(StorageKey::UnitShareCumulativeInfo),
-            wnear_id: None
+            wnear_id: None,
+            auto_whitelisted_postfix: HashSet::new()
         }
     }
 
@@ -217,7 +220,7 @@ impl Contract {
                 for token in action.tokens() {
                     assert!(
                         account.get_balance(&token).is_some() 
-                            || self.whitelisted_tokens.contains(&token),
+                            || self.is_whitelisted_token(&token),
                         "{}",
                         // [AUDIT_05]
                         ERR27_DEPOSIT_NEEDED
@@ -456,12 +459,25 @@ impl Contract {
     /// the async return of update_token_rate
     #[private]
     pub fn update_token_rate_callback(&mut self, token_id: AccountId) {
-        assert_eq!(env::promise_results_count(), 1, "{}", ERR123_ONE_PROMISE_RESULT);
-        let cross_call_result = match env::promise_result(0) {
-            PromiseResult::Successful(result) => result,
-            _ => env::panic(ERR124_CROSS_CALL_FAILED.as_bytes()),
+        let cross_call_result = if env::promise_results_count() == 1 {
+            let cross_call_result = match env::promise_result(0) {
+                PromiseResult::Successful(result) => result,
+                _ => env::panic(ERR124_CROSS_CALL_FAILED.as_bytes()),
+            };
+            cross_call_result
+        } else {
+            // Only SfraxRate + pyth
+            assert_eq!(env::promise_results_count(), 2, "{}", ERR123_TWO_PROMISE_RESULT);
+            let cross_call_result1 = match env::promise_result(0) {
+                PromiseResult::Successful(result) => result,
+                _ => env::panic(ERR124_CROSS_CALL_FAILED.as_bytes()),
+            };
+            let cross_call_result2 = match env::promise_result(1) {
+                PromiseResult::Successful(result) => result,
+                _ => env::panic(ERR124_CROSS_CALL_FAILED.as_bytes()),
+            };
+            pair_rated_price_to_vec_u8(cross_call_result1, cross_call_result2)
         };
-
         if let Some(mut rate) = global_get_rate(&token_id) {
             let new_rate = rate.set(&cross_call_result);
             global_set_rate(&token_id, &rate);
@@ -490,6 +506,10 @@ impl Contract {
         )
         .collect();
         assert_eq!(frozens.len(), 0, "{}", ERR52_FROZEN_TOKEN);
+    }
+
+    fn is_whitelisted_token(&self, token_id: &AccountId) -> bool {
+        self.whitelisted_tokens.contains(token_id) || self.auto_whitelisted_postfix.iter().any(|postfix| token_id.ends_with(postfix))
     }
 
     /// Check how much storage taken costs and refund the left over back.
@@ -612,7 +632,6 @@ impl Contract {
     }
 }
 
-use std::collections::HashMap;
 
 impl Contract {
     fn internal_execute_actions_by_cache(
@@ -1112,6 +1131,60 @@ mod tests {
             .predecessor_account_id(ValidAccountId::try_from("malicious").unwrap())
             .build());
         contract.ft_on_transfer(acc, U128(1_000), "".to_string());
+    }
+
+    #[test]
+    fn test_auto_whitelisted_postfix() {
+        let (mut context, mut contract) = setup_contract();
+        let acc = ValidAccountId::try_from("test_user").unwrap();
+        let token1 = ValidAccountId::try_from("test.abc.near").unwrap();
+        let token2 = ValidAccountId::try_from("test.def.near").unwrap();
+        testing_env!(context
+            .predecessor_account_id(acc.clone())
+            .attached_deposit(to_yocto("1"))
+            .build());
+        contract.storage_deposit(None, None);
+        testing_env!(context.predecessor_account_id(accounts(0)).attached_deposit(1).build());
+        contract.extend_auto_whitelisted_postfix(vec!["abc.near".to_string(), "def.near".to_string()]);
+        testing_env!(context
+            .predecessor_account_id(token1.clone())
+            .attached_deposit(1)
+            .build());
+        contract.ft_on_transfer(acc.clone(), U128(1_000_000), "".to_string());
+
+        testing_env!(context
+            .predecessor_account_id(token2.clone())
+            .attached_deposit(1)
+            .build());
+        contract.ft_on_transfer(acc.clone(), U128(1_000_000), "".to_string());
+
+        testing_env!(context
+            .predecessor_account_id(acc.clone())
+            .attached_deposit(env::storage_byte_cost() * 500)
+            .build());
+        let pool_id = contract.add_simple_pool(vec![token1.clone(), token2.clone()], 25);
+        testing_env!(context
+            .predecessor_account_id(acc.clone())
+            .attached_deposit(to_yocto("0.1"))
+            .build());
+        contract.add_liquidity(
+            pool_id,
+            vec![U128(10000), U128(10000)],
+            None,
+        );
+        
+        let actions: Vec<Action> = vec![Action::Swap(SwapAction{
+            pool_id,
+            token_in: token1.to_string(),
+            amount_in: Some(U128(10)),
+            token_out: token2.to_string(),
+            min_amount_out: U128(0),
+        })];
+        
+        testing_env!(context
+            .predecessor_account_id(acc.clone())
+            .build());
+        contract.execute_actions(actions, None);
     }
 
     #[test]